Avoid casting of 1/n to int

+++ src/gosdt.cpp
@@ -50,11 +50,11 @@ void GOSDT::fit(std::istream & data_source, std::unordered_set< Model > & models
     optimizer.load(data_source);
 
     int const n = State::dataset.size();
-    if(Configuration::regularization < 1/n) {
+    if(Configuration::regularization < (float) 1/n) {
         std::cout << "Regularization smaller than 1/(num_samples) - this may lead to longer training time if not adjusted." << std::endl;
         if (!Configuration::allow_small_reg) {
             std::cout << "Regularization increased to 1/(num_samples) = " << (1/n) << ". To allow regularization below this, set allow_small_reg to true" << std::endl;
-            Configuration::regularization = 1/n;
+            Configuration::regularization = (float) 1/n;
         }
     }
